add Logger struct
